src/third_party/blink/web_tests/virtual/android/fullscreen/video-fail-to-enter-full-screen-expected.txt

CONSOLE WARNING: line 12: Failed to execute 'requestFullscreen' on 'Element': API can only be initiated by a user gesture.
